Fix a Hostmot2 Resolver bug in 64 bit builds. Do not assume that "long" is 32 bits in all systems. this bug resulted in the resolver module being utterly unusable, with the apparent resolver angle varying randomly by thousands of turns

+++ src/hal/drivers/mesa-hostmot2/hostmot2.h
@@ -345,9 +345,9 @@ typedef struct {
 
     } hal;
     
-    long long accum;
-    long long offset;
-    u32 old_reg;
+    __s64 accum;
+    __s64 offset;
+    __u32 old_reg;
 
 } hm2_resolver_instance_t;
